PSIQJE-7 PSIQJE-8 Mise en place de la méthode onReceive pour la classe ParkingLot et ParkingSensor + ajout/modification des messages InitParking + SendRequest + StartParkingSimulation

+++ src/main/java/smart/campus/simulation/messages/StartParkingSimulation.java
@@ -4,9 +4,9 @@ public class StartParkingSimulation {
 	private final int begin;
 	private final int duration;
 	private final int interval;
-	private final int value;
+	private final float value;
 	
-	public StartParkingSimulation(int b, int d, int i, int v) {
+	public StartParkingSimulation(int b, int d, int i, float v) {
 		this.begin = b;
 		this.duration = d;
 		this.interval = i;
@@ -25,7 +25,7 @@ public class StartParkingSimulation {
 		return interval;
 	}
 
-	public int getValue() {
+	public float getValue() {
 		return value;
 	}
 }
